+++ lib/io.dart
@@ -5,6 +5,6 @@
 export 'src/copy_path.dart' show copyPath, copyPathSync;
 export 'src/exit_code.dart' show ExitCode;
 export 'src/permissions.dart' show isExecutable;
-export 'src/process_manager.dart' show ProcessManager, Spawn;
+export 'src/process_manager.dart' show ProcessManager, Spawn, StartProcess;
 export 'src/shared_stdin.dart' show SharedStdIn, sharedStdIn;
 export 'src/shell_words.dart' show shellSplit;